# PERF: Styler render improvement by limiting .join()