--- a/pandas/io/formats/templates/html.tpl
+++ b/pandas/io/formats/templates/html.tpl
@@ -39,7 +39,7 @@
     <tr>
 {% for c in r %}
 {% if c.is_visible != False %}
-      <{{c.type}} class="{{c.class}}" {{c.attributes|join(" ")}}>{{c.value}}</{{c.type}}>
+      <{{c.type}} class="{{c.class}}" {{c.attributes}}>{{c.value}}</{{c.type}}>
 {% endif %}
 {% endfor %}
     </tr>
@@ -56,11 +56,7 @@
     <tr>
 {% for c in r %}
 {% if c.is_visible != False %}
-{% if c.type == 'th' %}
-      <th {% if c.id is defined -%} id="T_{{uuid}}{{c.id}}" {%- endif %} class="{{c.class}}" {{c.attributes|join(" ")}}>{{c.display_value}}</th>
-{% else %}
-      <td {% if c.id is defined -%} id="T_{{uuid}}{{c.id}}" {%- endif %} class="{{c.class}}" >{{c.display_value}}</td>
-{% endif %}
+      <{{c.type}} {% if c.id is defined -%} id="T_{{uuid}}{{c.id}}" {%- endif %} class="{{c.class}}" {{c.attributes}}>{{c.display_value}}</{{c.type}}>
 {% endif %}
 {% endfor %}
     </tr>